# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Notes App API",
    "description": "A multi-user notes service with JWT auth, sharing, pinning + reordering, full-text search, and pagination.",
    "version": "1.0.0",
    "contact": {
      "name": "Anmol Singh",
      "email": "anmol.singh@withinstafix.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Note": {
        "type": "object",
        "required": ["id", "title", "content", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string", "maxLength": 255 },
          "content": { "type": "string", "maxLength": 100000 },
          "pinned": { "type": "boolean" },
          "position": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "Error": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string" },
          "issues": {
            "type": "object",
            "additionalProperties": { "type": "array", "items": { "type": "string" } }
          },
          "errors": { "type": "array", "items": { "type": "string" } }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 8, "maxLength": 128 }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": ["access_token"],
        "properties": {
          "access_token": { "type": "string" }
        }
      },
      "CreateNoteRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 255 },
          "content": { "type": "string", "maxLength": 100000 }
        }
      },
      "UpdateNoteRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 255 },
          "content": { "type": "string", "maxLength": 100000 }
        },
        "description": "At least one of title or content must be provided."
      },
      "ShareNoteRequest": {
        "type": "object",
        "required": ["share_with_email"],
        "properties": {
          "share_with_email": { "type": "string", "format": "email" }
        }
      },
      "PinNoteRequest": {
        "type": "object",
        "required": ["pinned"],
        "properties": {
          "pinned": { "type": "boolean" }
        }
      },
      "NoteVersion": {
        "type": "object",
        "required": ["id", "version_no", "title", "content", "created_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "version_no": { "type": "integer", "minimum": 1 },
          "title": { "type": "string" },
          "content": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "ReorderNotesRequest": {
        "type": "object",
        "required": ["note_ids"],
        "properties": {
          "note_ids": {
            "type": "array",
            "minItems": 1,
            "maxItems": 1000,
            "items": { "type": "string", "format": "uuid" }
          }
        }
      },
      "AboutResponse": {
        "type": "object",
        "required": ["name", "email", "my features"],
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "my features": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          }
        }
      }
    },
    "parameters": {
      "NoteId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      },
      "Offset": {
        "name": "offset",
        "in": "query",
        "schema": { "type": "integer", "minimum": 0, "default": 0 }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "BadRequest": {
        "description": "Validation failed",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "Resource not found or not accessible",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Conflict": {
        "description": "Conflict (e.g. duplicate email)",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "security": [],
        "responses": {
          "200": {
            "description": "Server is up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "status": { "type": "string", "example": "ok" } }
                }
              }
            }
          }
        }
      }
    },
    "/about": {
      "get": {
        "summary": "Author + chosen feature info",
        "security": [],
        "responses": {
          "200": {
            "description": "About info",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AboutResponse" } }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "This OpenAPI document",
        "security": [],
        "responses": {
          "200": {
            "description": "OpenAPI 3.0 spec",
            "content": { "application/json": {} }
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Register a new user",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "User registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Authenticate and receive a JWT",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": {
            "description": "Invalid email or password",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/notes": {
      "get": {
        "summary": "List notes owned by or shared with the user",
        "parameters": [
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Offset" }
        ],
        "responses": {
          "200": {
            "description": "Notes list (X-Total-Count header carries total)",
            "headers": {
              "X-Total-Count": {
                "description": "Total number of notes accessible to the user",
                "schema": { "type": "integer" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Note" } }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "summary": "Create a new note",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateNoteRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "Note created",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Note" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/notes/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/NoteId" }],
      "get": {
        "summary": "Fetch a single note (owner or shared user)",
        "responses": {
          "200": {
            "description": "The note",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Note" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "summary": "Update a note (owner only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateNoteRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Updated note",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Note" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "summary": "Delete a note (owner only)",
        "responses": {
          "204": { "description": "Deleted" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/notes/{id}/versions": {
      "parameters": [{ "$ref": "#/components/parameters/NoteId" }],
      "get": {
        "summary": "List version history for a note (owner or shared user)",
        "responses": {
          "200": {
            "description": "Versions ordered newest first",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/NoteVersion" } }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/notes/{id}/versions/{versionId}": {
      "parameters": [
        { "$ref": "#/components/parameters/NoteId" },
        { "name": "versionId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
      ],
      "get": {
        "summary": "Fetch a single version (owner or shared user)",
        "responses": {
          "200": {
            "description": "The version snapshot",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NoteVersion" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/notes/{id}/versions/{versionId}/restore": {
      "parameters": [
        { "$ref": "#/components/parameters/NoteId" },
        { "name": "versionId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
      ],
      "post": {
        "summary": "Restore a note to a previous version (owner only). Creates a new version with the restored content.",
        "responses": {
          "200": {
            "description": "Note after restore",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Note" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/notes/{id}/share": {
      "parameters": [{ "$ref": "#/components/parameters/NoteId" }],
      "post": {
        "summary": "Share a note with another user by email (owner only, idempotent)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ShareNoteRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Note shared (or already shared)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/notes/{id}/pin": {
      "parameters": [{ "$ref": "#/components/parameters/NoteId" }],
      "put": {
        "summary": "Pin or unpin a note (owner only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PinNoteRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Pin status updated",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Note" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/notes/reorder": {
      "put": {
        "summary": "Reorder notes by setting positions in array order (owner-only IDs are updated, others silently skipped)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ReorderNotesRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Reorder result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "count": { "type": "integer", "description": "Number of notes actually updated" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/search": {
      "get": {
        "summary": "Case-insensitive search across title and content of accessible notes",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "maxLength": 255 }
          },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Offset" }
        ],
        "responses": {
          "200": {
            "description": "Matching notes (X-Total-Count header carries total)",
            "headers": {
              "X-Total-Count": {
                "description": "Total number of matching notes",
                "schema": { "type": "integer" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Note" } }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    }
  }
}